Front End Complete
Finished up the UI and added classes that can be handled by the
script. The "x" and "o" are created completely within the CSS and
applied by adding a .x or .o class to a cell div. Hover affect is
achieved by adding .x or .o to the board div.

HTML and CSS files are finished.
Next is the JS script.